Previous to Controller
Avance/corrección de clases/interfaces de Nutritionist

+++ src/main/java/com/appnutricare/service/CrudService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface CrudService<T> {
     T save(T t) throws Exception;
-    void delete(Long id) throws Exception;
+    void delete(int id) throws Exception;
     List<T> getAll() throws  Exception;
-    Optional<T> getById(Long id) throws Exception;
+    Optional<T> getById(int id) throws Exception;
 }
